requirements.txt
# Core Dependencies
crewai>=0.11.0
langchain>=0.1.0
openai>=1.3.7
streamlit>=1.29.0
pydantic>=2.5.0
PyYAML>=6.0.1

# Streamlit Components
streamlit-option-menu>=0.3.6
streamlit-image-select>=0.6.0
streamlit-camera-input-live>=0.2.0
streamlit-extras>=0.3.5
streamlit-autorefresh>=1.0.1
streamlit-custom-notification-box>=0.1.1
streamlit-toggle-switch>=1.0.2
streamlit-card>=1.0.0
streamlit-faker>=0.0.3
hydralit_components>=1.0.10
extra-streamlit-components>=0.1.60

# UI/UX
plotly>=5.18.0
altair>=5.2.0
bokeh>=3.3.1
streamlit-lottie>=0.0.5
streamlit-authenticator>=0.2.3

# Export
fpdf2>=2.7.6
openpyxl>=3.1.2
XlsxWriter>=3.1.9
python-docx>=1.1.0
pdfkit>=1.0.0
reportlab>=4.0.8

# Image Processing
opencv-python>=4.8.0
Pillow>=10.1.0
pytesseract>=0.3.10

# Diagnostic & CAN Bus
python-can>=4.2.2
cantools>=39.2.0
obd>=0.7.1

# IoT & Telemetry
paho-mqtt>=1.6.1
influxdb-client>=1.38.0
pyserial>=3.5

# Video Processing
youtube-transcript-api>=0.6.1
moviepy>=1.0.3

# Database
supabase>=2.0.3
sqlalchemy>=2.0.23

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
pandas>=2.1.4
numpy>=1.26.2
pyyaml>=6.0.1
python-jose>=3.3.0
httpx>=0.25.2
markdown>=3.5.1
jinja2>=3.1.2
watchdog>=3.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.14
mkdocstrings>=0.23.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1